+++ src/main/java/org/c4marathon/assignment/controller/SettlementController.java
@@ -3,8 +3,8 @@ package org.c4marathon.assignment.controller;
 import java.util.List;
 
 import org.c4marathon.assignment.common.response.ApiResponse;
-import org.c4marathon.assignment.dto.sattlement.SettlementRequestDto;
-import org.c4marathon.assignment.dto.sattlement.SettlementResponseDto;
+import org.c4marathon.assignment.dto.settlement.SettlementRequestDto;
+import org.c4marathon.assignment.dto.settlement.SettlementResponseDto;
 import org.c4marathon.assignment.usecase.SettlementUseCase;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
