+++ src/main/java/org/c4marathon/assignment/dto/sattlement/SettlementResponseDto.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.dto.sattlement;
+package org.c4marathon.assignment.dto.settlement;
 
 import jakarta.validation.constraints.NotNull;
