+++ src/main/java/org/c4marathon/assignment/usecase/SettlementUseCase.java
@@ -3,8 +3,8 @@ package org.c4marathon.assignment.usecase;
 import java.util.List;
 
 import org.c4marathon.assignment.domain.service.mainaccount.SettlementService;
-import org.c4marathon.assignment.dto.sattlement.SettlementRequestDto;
-import org.c4marathon.assignment.dto.sattlement.SettlementResponseDto;
+import org.c4marathon.assignment.dto.settlement.SettlementRequestDto;
+import org.c4marathon.assignment.dto.settlement.SettlementResponseDto;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
